+++ src/application/style/thema.rs
@@ -16,8 +16,8 @@ pub enum Thema {
     /// Die helle Variante.
     #[default]
     Hell,
-    // /// Die dunkle Variante.
-    // Dunkel,
+    /// Die dunkle Variante.
+    Dunkel,
 }
 
 impl Thema {
@@ -26,6 +26,7 @@ impl Thema {
     pub fn strich(&self) -> Farbe {
         match self {
             Thema::Hell => Farbe { rot: 0., grün: 0., blau: 0. },
+            Thema::Dunkel => Farbe { rot: 1., grün: 1., blau: 1. },
         }
     }
 
@@ -36,6 +37,9 @@ impl Thema {
             Thema::Hell if aktiv => 0.5,
             Thema::Hell if in_bounds => 0.7,
             Thema::Hell => 0.9,
+            Thema::Dunkel if aktiv => 0.5,
+            Thema::Dunkel if in_bounds => 0.3,
+            Thema::Dunkel => 0.5,
         };
         Farbe { rot: grey_value, grün: grey_value, blau: grey_value }
     }
@@ -47,6 +51,7 @@ impl StyleSheet for Thema {
     fn appearance(&self, style: &Self::Style) -> Appearance {
         match self {
             Thema::Hell => StyleSheet::appearance(&Theme::Light, style),
+            Thema::Dunkel => StyleSheet::appearance(&Theme::Dark, style),
         }
     }
 }
@@ -57,12 +62,14 @@ impl checkbox::StyleSheet for Thema {
     fn active(&self, style: &Self::Style, is_checked: bool) -> checkbox::Appearance {
         match self {
             Thema::Hell => checkbox::StyleSheet::active(&Theme::Light, style, is_checked),
+            Thema::Dunkel => checkbox::StyleSheet::active(&Theme::Dark, style, is_checked),
         }
     }
 
     fn hovered(&self, style: &Self::Style, is_checked: bool) -> checkbox::Appearance {
         match self {
             Thema::Hell => checkbox::StyleSheet::hovered(&Theme::Light, style, is_checked),
+            Thema::Dunkel => checkbox::StyleSheet::hovered(&Theme::Dark, style, is_checked),
         }
     }
 }
@@ -73,6 +80,7 @@ impl text::StyleSheet for Thema {
     fn appearance(&self, style: Self::Style) -> text::Appearance {
         match self {
             Thema::Hell => text::StyleSheet::appearance(&Theme::Light, style),
+            Thema::Dunkel => text::StyleSheet::appearance(&Theme::Dark, style),
         }
     }
 }
@@ -83,48 +91,56 @@ impl text_input::StyleSheet for Thema {
     fn active(&self, style: &Self::Style) -> text_input::Appearance {
         match self {
             Thema::Hell => text_input::StyleSheet::active(&Theme::Light, style),
+            Thema::Dunkel => text_input::StyleSheet::active(&Theme::Dark, style),
         }
     }
 
     fn focused(&self, style: &Self::Style) -> text_input::Appearance {
         match self {
             Thema::Hell => text_input::StyleSheet::focused(&Theme::Light, style),
+            Thema::Dunkel => text_input::StyleSheet::focused(&Theme::Dark, style),
         }
     }
 
     fn placeholder_color(&self, style: &Self::Style) -> iced_core::Color {
         match self {
             Thema::Hell => text_input::StyleSheet::placeholder_color(&Theme::Light, style),
+            Thema::Dunkel => text_input::StyleSheet::placeholder_color(&Theme::Dark, style),
         }
     }
 
     fn value_color(&self, style: &Self::Style) -> iced_core::Color {
         match self {
             Thema::Hell => text_input::StyleSheet::value_color(&Theme::Light, style),
+            Thema::Dunkel => text_input::StyleSheet::value_color(&Theme::Dark, style),
         }
     }
 
     fn disabled_color(&self, style: &Self::Style) -> iced_core::Color {
         match self {
             Thema::Hell => text_input::StyleSheet::disabled_color(&Theme::Light, style),
+            Thema::Dunkel => text_input::StyleSheet::disabled_color(&Theme::Dark, style),
         }
     }
 
     fn selection_color(&self, style: &Self::Style) -> iced_core::Color {
         match self {
             Thema::Hell => text_input::StyleSheet::selection_color(&Theme::Light, style),
+            Thema::Dunkel => text_input::StyleSheet::selection_color(&Theme::Dark, style),
         }
     }
 
     fn disabled(&self, style: &Self::Style) -> text_input::Appearance {
         match self {
             Thema::Hell => text_input::StyleSheet::disabled(&Theme::Light, style),
+            Thema::Dunkel => text_input::StyleSheet::disabled(&Theme::Dark, style),
         }
     }
 
     fn hovered(&self, style: &Self::Style) -> text_input::Appearance {
         match self {
             Thema::Hell => text_input::StyleSheet::hovered(&Theme::Light, style),
+            Thema::Dunkel => text_input::StyleSheet::hovered(&Theme::Dark, style),
         }
     }
 }
@@ -135,6 +151,7 @@ impl card::StyleSheet for Thema {
     fn active(&self, style: &Self::Style) -> card::Appearance {
         match self {
             Thema::Hell => card::StyleSheet::active(&Theme::Light, style),
+            Thema::Dunkel => card::StyleSheet::active(&Theme::Dark, style),
         }
     }
 }
@@ -145,18 +162,21 @@ impl slider::StyleSheet for Thema {
     fn active(&self, style: &Self::Style) -> vertical_slider::Appearance {
         match self {
             Thema::Hell => slider::StyleSheet::active(&Theme::Light, style),
+            Thema::Dunkel => slider::StyleSheet::active(&Theme::Dark, style),
         }
     }
 
     fn hovered(&self, style: &Self::Style) -> vertical_slider::Appearance {
         match self {
             Thema::Hell => slider::StyleSheet::hovered(&Theme::Light, style),
+            Thema::Dunkel => slider::StyleSheet::hovered(&Theme::Dark, style),
         }
     }
 
     fn dragging(&self, style: &Self::Style) -> vertical_slider::Appearance {
         match self {
             Thema::Hell => slider::StyleSheet::dragging(&Theme::Light, style),
+            Thema::Dunkel => slider::StyleSheet::dragging(&Theme::Dark, style),
         }
     }
 }
@@ -167,12 +187,14 @@ impl radio::StyleSheet for Thema {
     fn active(&self, style: &Self::Style, is_selected: bool) -> radio::Appearance {
         match self {
             Thema::Hell => radio::StyleSheet::active(&Theme::Light, style, is_selected),
+            Thema::Dunkel => radio::StyleSheet::active(&Theme::Dark, style, is_selected),
         }
     }
 
     fn hovered(&self, style: &Self::Style, is_selected: bool) -> radio::Appearance {
         match self {
             Thema::Hell => radio::StyleSheet::hovered(&Theme::Light, style, is_selected),
+            Thema::Dunkel => radio::StyleSheet::hovered(&Theme::Dark, style, is_selected),
         }
     }
 }
@@ -183,18 +205,21 @@ impl number_input::StyleSheet for Thema {
     fn active(&self, style: &Self::Style) -> number_input::Appearance {
         match self {
             Thema::Hell => number_input::StyleSheet::active(&Theme::Light, style),
+            Thema::Dunkel => number_input::StyleSheet::active(&Theme::Dark, style),
         }
     }
 
     fn pressed(&self, style: &Self::Style) -> number_input::Appearance {
         match self {
             Thema::Hell => number_input::StyleSheet::pressed(&Theme::Light, style),
+            Thema::Dunkel => number_input::StyleSheet::pressed(&Theme::Dark, style),
         }
     }
 
     fn disabled(&self, style: &Self::Style) -> number_input::Appearance {
         match self {
             Thema::Hell => number_input::StyleSheet::disabled(&Theme::Light, style),
+            Thema::Dunkel => number_input::StyleSheet::disabled(&Theme::Dark, style),
         }
     }
 }