+++ ChangeLog.md
@@ -2,6 +2,9 @@
 
 ## Unreleased changes
 
+- Aufteilen in mehrere crates.
+- Dunkles Thema.
+
 # 4.0.3
 
 - Knöpfe zeigen wieder an, wenn sich der Mauszeiger über ihnen befindet.